refactor: move cleanExpiredCache to own module
